+++ extension/src/popup/views/SignBlob/styles.scss
@@ -1,4 +1,4 @@
-.SignTransaction {
+.SignBlob {
   height: var(--popup--height);
   overflow: hidden;
   position: relative;
