+++ extension/src/popup/views/SignTransaction/index.tsx
@@ -1,10 +1,15 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import { useLocation } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { getTransactionInfo } from "helpers/stellar";
-import { rejectTransaction, rejectBlob } from "popup/ducks/access";
+import { useTranslation, Trans } from "react-i18next";
+import { Card, Icon } from "@stellar/design-system";
+import * as SorobanSdk from "soroban-client";
+import StellarSdk, { FederationServer, MuxedAccount } from "stellar-sdk";
+
+import { signTransaction, rejectTransaction } from "popup/ducks/access";
 import {
   allAccountsSelector,
+  confirmPassword,
   hasPrivateKeySelector,
   makeAccountActive,
   publicKeySelector,
@@ -16,28 +21,65 @@ import {
 } from "popup/ducks/settings";
 
 import {
+  startHwSign,
   ShowOverlayStatus,
   transactionSubmissionSelector,
 } from "popup/ducks/transactionSubmission";
+
+import { TRANSACTION_WARNING } from "constants/transaction";
+
+import { emitMetric } from "helpers/metrics";
+import {
+  getTransactionInfo,
+  isFederationAddress,
+  isMuxedAccount,
+  truncatedPublicKey,
+} from "helpers/stellar";
+import { decodeMemo } from "popup/helpers/parseTransaction";
+import { Button } from "popup/basics/buttons/Button";
+import { InfoBlock } from "popup/basics/InfoBlock";
+import { TransactionHeading } from "popup/basics/TransactionHeading";
+
+import {
+  ButtonsContainer,
+  ModalHeader,
+  ModalWrapper,
+} from "popup/basics/Modal";
+
+import { METRIC_NAMES } from "popup/constants/metricsNames";
+
+import { AccountListIdenticon } from "popup/components/identicons/AccountListIdenticon";
+import { AccountList, OptionTag } from "popup/components/account/AccountList";
+import { PunycodedDomain } from "popup/components/PunycodedDomain";
+import {
+  WarningMessageVariant,
+  WarningMessage,
+  FirstTimeWarningMessage,
+  FlaggedWarningMessage,
+} from "popup/components/WarningMessages";
+import { Transaction } from "popup/components/signTransaction/Transaction";
+import { LedgerSign } from "popup/components/hardwareConnect/LedgerSign";
+import { SlideupModal } from "popup/components/SlideupModal";
+
+import { VerifyAccount } from "popup/views/VerifyAccount";
 
 import { Account } from "@shared/api/types";
 import { AppDispatch } from "popup/App";
 
 import "./styles.scss";
-import { SignBlobBody } from "./SignBlob";
-import { SignTxBody } from "./SignTx";
+
+import { FlaggedKeys } from "types/transactions";
+import { TransactionInfo } from "popup/components/signTransaction/TransactionInfo";
 
 export const SignTransaction = () => {
   const location = useLocation();
-  const blobOrTx = getTransactionInfo(location.search);
-  const isBlob = "blob" in blobOrTx;
-
+  const { t } = useTranslation();
+  const isExperimentalModeEnabled = useSelector(
+    settingsExperimentalModeSelector,
+  );
   const dispatch: AppDispatch = useDispatch();
   const { networkName, networkPassphrase } = useSelector(
     settingsNetworkDetailsSelector,
-  );
-  const isExperimentalModeEnabled = useSelector(
-    settingsExperimentalModeSelector,
   );
 
   const hardwareWalletType = useSelector(hardwareWalletTypeSelector);
@@ -60,11 +102,45 @@ export const SignTransaction = () => {
   // the public key the user had selected before starting this flow
   const defaultPublicKey = useRef(publicKey);
   const allAccountsMap = useRef({} as { [key: string]: Account });
-  const accountToSign = blobOrTx.accountToSign; // both types have this key
+
+  const tx = getTransactionInfo(location.search);
+
+  const {
+    accountToSign: _accountToSign,
+    transactionXdr,
+    domain,
+    isDomainListedAllowed,
+    isHttpsDomain,
+    flaggedKeys,
+  } = tx;
+
+  /* 
+  Reconstruct the tx from xdr as passing a tx through extension contexts 
+  loses custom prototypes associated with some values. This is fine for most cases 
+  where we just need a high level overview of the tx, like just a list of operations.  
+  But in this case, we will need the hostFn prototype associated with Soroban tx operations.
+  */
+
+  const SDK = isExperimentalModeEnabled ? SorobanSdk : StellarSdk;
+  const transaction = SDK.TransactionBuilder.fromXDR(
+    transactionXdr,
+    networkPassphrase,
+  );
+
+  const {
+    _fee,
+    _innerTransaction,
+    _memo,
+    _networkPassphrase,
+    _sequence,
+  } = transaction;
+
+  const isFeeBump = !!_innerTransaction;
+  const memo = decodeMemo(_memo);
+  let accountToSign = _accountToSign;
 
   const rejectAndClose = () => {
-    const _reject = isBlob ? rejectTransaction : rejectBlob;
-    dispatch(_reject());
+    dispatch(rejectTransaction());
     window.close();
   };
 
@@ -118,52 +194,266 @@ export const SignTransaction = () => {
     setCurrentAccount(allAccountsMap.current[publicKey] || ({} as Account));
   }, [allAccounts, publicKey]);
 
-  if ("blob" in blobOrTx) {
+  useEffect(() => {
+    if (startedHwSign && hwStatus === ShowOverlayStatus.IDLE) {
+      window.close();
+    }
+  }, [startedHwSign, hwStatus]);
+
+  const signAndClose = async () => {
+    if (isHardwareWallet) {
+      await dispatch(
+        startHwSign({ transactionXDR: transactionXdr, shouldSubmit: false }),
+      );
+      setStartedHwSign(true);
+    } else {
+      await dispatch(signTransaction());
+      window.close();
+    }
+  };
+
+  const _handleApprove = handleApprove(signAndClose);
+
+  const verifyPasswordThenSign = async (password: string) => {
+    const confirmPasswordResp = await dispatch(confirmPassword(password));
+
+    if (confirmPassword.fulfilled.match(confirmPasswordResp)) {
+      await signAndClose();
+    }
+  };
+
+  const flaggedKeyValues = Object.values(flaggedKeys as FlaggedKeys);
+  const isUnsafe = flaggedKeyValues.some(({ tags }) =>
+    tags.includes(TRANSACTION_WARNING.unsafe),
+  );
+  const isMalicious = flaggedKeyValues.some(({ tags }) =>
+    tags.includes(TRANSACTION_WARNING.malicious),
+  );
+  const isMemoRequired = flaggedKeyValues.some(
+    ({ tags }) => tags.includes(TRANSACTION_WARNING.memoRequired) && !memo,
+  );
+
+  const resolveFederatedAddress = useCallback(async (inputDest) => {
+    let resolvedPublicKey;
+    try {
+      const fedResp = await FederationServer.resolve(inputDest);
+      resolvedPublicKey = fedResp.account_id;
+    } catch (e) {
+      console.error(e);
+    }
+
+    return resolvedPublicKey;
+  }, []);
+
+  const decodeAccountToSign = async () => {
+    if (_accountToSign) {
+      if (isMuxedAccount(_accountToSign)) {
+        const mAccount = MuxedAccount.fromAddress(_accountToSign, "0");
+        accountToSign = mAccount.baseAccount().accountId();
+      }
+      if (isFederationAddress(_accountToSign)) {
+        accountToSign = (await resolveFederatedAddress(
+          accountToSign,
+        )) as string;
+      }
+    }
+  };
+  decodeAccountToSign();
+
+  useEffect(() => {
+    if (isMemoRequired) {
+      emitMetric(METRIC_NAMES.signTransactionMemoRequired);
+    }
+    if (isUnsafe) {
+      emitMetric(METRIC_NAMES.signTransactionUnsafe);
+    }
+    if (isMalicious) {
+      emitMetric(METRIC_NAMES.signTransactionMalicious);
+    }
+  }, [isMemoRequired, isMalicious, isUnsafe]);
+
+  useEffect(() => {
+    // handle any changes to the current acct - whether by auto select or manual select
+    setCurrentAccount(allAccountsMap.current[publicKey] || ({} as Account));
+  }, [allAccountsMap, allAccounts, publicKey, setCurrentAccount]);
+
+  const isSubmitDisabled = isMemoRequired || isMalicious;
+
+  if (_networkPassphrase !== networkPassphrase) {
     return (
-      <SignBlobBody
-        accountNotFound={accountNotFound}
-        allAccounts={allAccounts}
-        blob={blobOrTx}
-        currentAccount={currentAccount}
-        handleApprove={handleApprove}
-        hwStatus={hwStatus}
-        isConfirming={isConfirming}
-        isDropdownOpen={isDropdownOpen}
-        isExperimentalModeEnabled={isExperimentalModeEnabled}
-        isHardwareWallet={isHardwareWallet}
-        isPasswordRequired={isPasswordRequired}
-        publicKey={publicKey}
-        rejectAndClose={rejectAndClose}
-        setIsDropdownOpen={setIsDropdownOpen}
-        setIsPasswordRequired={setIsPasswordRequired}
-        setStartedHwSign={setStartedHwSign}
-      />
+      <ModalWrapper>
+        <WarningMessage
+          variant={WarningMessageVariant.warning}
+          handleCloseClick={() => window.close()}
+          isActive
+          header={`${t("Freighter is set to")} ${networkName}`}
+        >
+          <p>
+            {t("The transaction you’re trying to sign is on")}{" "}
+            {_networkPassphrase}.
+          </p>
+          <p>{t("Signing this transaction is not possible at the moment.")}</p>
+        </WarningMessage>
+      </ModalWrapper>
     );
   }
 
-  return (
-    <SignTxBody
-      accountNotFound={accountNotFound}
-      allAccounts={allAccounts}
-      allAccountsMap={allAccountsMap}
-      currentAccount={currentAccount}
-      handleApprove={handleApprove}
-      hwStatus={hwStatus}
-      isConfirming={isConfirming}
-      isDropdownOpen={isDropdownOpen}
-      isExperimentalModeEnabled={isExperimentalModeEnabled}
-      isHardwareWallet={isHardwareWallet}
-      isPasswordRequired={isPasswordRequired}
-      networkName={networkName}
-      networkPassphrase={networkPassphrase}
-      publicKey={publicKey}
-      rejectAndClose={rejectAndClose}
-      setCurrentAccount={setCurrentAccount}
-      setIsDropdownOpen={setIsDropdownOpen}
-      setIsPasswordRequired={setIsPasswordRequired}
-      setStartedHwSign={setStartedHwSign}
-      startedHwSign={startedHwSign}
-      tx={blobOrTx as any} // TODO
+  if (!isHttpsDomain && !isExperimentalModeEnabled) {
+    return (
+      <ModalWrapper>
+        <WarningMessage
+          handleCloseClick={() => window.close()}
+          isActive
+          variant={WarningMessageVariant.warning}
+          header={t("WEBSITE CONNECTION IS NOT SECURE")}
+        >
+          <p>
+            <Trans domain={domain}>
+              The website <strong>{{ domain }}</strong> does not use an SSL
+              certificate. For additional safety Freighter only works with
+              websites that provide an SSL certificate.
+            </Trans>
+          </p>
+        </WarningMessage>
+      </ModalWrapper>
+    );
+  }
+
+  return isPasswordRequired ? (
+    <VerifyAccount
+      isApproval
+      customBackAction={() => setIsPasswordRequired(false)}
+      customSubmit={verifyPasswordThenSign}
     />
+  ) : (
+    <>
+      {hwStatus === ShowOverlayStatus.IN_PROGRESS && <LedgerSign />}
+      <div className="SignTransaction" data-testid="SignTransaction">
+        <ModalWrapper>
+          <ModalHeader>
+            <strong>{t("Confirm Transaction")}</strong>
+          </ModalHeader>
+          {isExperimentalModeEnabled ? (
+            <WarningMessage
+              header="Experimental Mode"
+              variant={WarningMessageVariant.default}
+            >
+              <p>
+                {t(
+                  "You are interacting with a transaction that may be using untested and changing schemas. Proceed at your own risk.",
+                )}
+              </p>
+            </WarningMessage>
+          ) : null}
+          {flaggedKeyValues.length ? (
+            <FlaggedWarningMessage
+              isUnsafe={isUnsafe}
+              isMalicious={isMalicious}
+              isMemoRequired={isMemoRequired}
+            />
+          ) : null}
+          {!isDomainListedAllowed && !isSubmitDisabled ? (
+            <FirstTimeWarningMessage />
+          ) : null}
+          <div className="SignTransaction__info">
+            <Card variant={Card.variant.highlight}>
+              <PunycodedDomain domain={domain} isRow />
+              <div className="SignTransaction__subject">
+                {t("is requesting approval to a")}{" "}
+                {isFeeBump ? "fee bump " : ""}
+                {t("transaction")}:
+              </div>
+              <div className="SignTransaction__approval">
+                <div className="SignTransaction__approval__title">
+                  {t("Approve using")}:
+                </div>
+                <div
+                  className="SignTransaction__current-account"
+                  onClick={() => setIsDropdownOpen(true)}
+                >
+                  <AccountListIdenticon
+                    displayKey
+                    accountName={currentAccount.name}
+                    active
+                    publicKey={currentAccount.publicKey}
+                    setIsDropdownOpen={setIsDropdownOpen}
+                  >
+                    <OptionTag
+                      hardwareWalletType={currentAccount.hardwareWalletType}
+                      imported={currentAccount.imported}
+                    />
+                  </AccountListIdenticon>
+                  <div className="SignTransaction__current-account__chevron">
+                    <Icon.ChevronDown />
+                  </div>
+                </div>
+              </div>
+            </Card>
+            {accountNotFound && accountToSign ? (
+              <div className="SignTransaction__account-not-found">
+                <InfoBlock variant={InfoBlock.variant.warning}>
+                  {t("The application is requesting a specific account")} (
+                  {truncatedPublicKey(accountToSign)}),{" "}
+                  {t(
+                    "which is not available on Freighter. If you own this account, you can import it into Freighter to complete this transaction.",
+                  )}
+                </InfoBlock>
+              </div>
+            ) : null}
+          </div>
+          {isFeeBump ? (
+            <div className="SignTransaction__inner-transaction">
+              <Transaction
+                flaggedKeys={flaggedKeys}
+                isMemoRequired={isMemoRequired}
+                transaction={_innerTransaction}
+              />
+            </div>
+          ) : (
+            <Transaction
+              flaggedKeys={flaggedKeys}
+              isMemoRequired={isMemoRequired}
+              transaction={transaction}
+            />
+          )}
+          <TransactionHeading>{t("Transaction Info")}</TransactionHeading>
+          <TransactionInfo
+            _fee={_fee}
+            _sequence={_sequence}
+            isFeeBump={isFeeBump}
+            isMemoRequired={isMemoRequired}
+          />
+        </ModalWrapper>
+        <ButtonsContainer>
+          <Button
+            fullWidth
+            variant={Button.variant.tertiary}
+            onClick={() => rejectAndClose()}
+          >
+            {t("Reject")}
+          </Button>
+          <Button
+            disabled={isSubmitDisabled}
+            fullWidth
+            isLoading={isConfirming}
+            onClick={() => _handleApprove()}
+          >
+            {t("Approve")}
+          </Button>
+        </ButtonsContainer>
+        <SlideupModal
+          isModalOpen={isDropdownOpen}
+          setIsModalOpen={setIsDropdownOpen}
+        >
+          <div className="SignTransaction__modal">
+            <AccountList
+              allAccounts={allAccounts}
+              publicKey={publicKey}
+              setIsDropdownOpen={setIsDropdownOpen}
+            />
+          </div>
+        </SlideupModal>
+      </div>
+    </>
   );
 };
